% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj-or.R
\name{adjusted.odds.ratio}
\alias{adjusted.odds.ratio}
\title{Adjusted odds ratio accounting for misclassification}
\usage{
adjusted.odds.ratio(data, sensitivity, specificity)
}
\arguments{
\item{data}{a 2 x 2 matrix with disease=yes in row 1 and exposure=yes in
column 1.}

\item{sensitivity}{proportion in interval (0,1) of exposed cohort who truly develop the disease
and are diagnosed correctly.}

\item{specificity}{proportion in interval (0,1) of exposed cohort who truly do not develop the
disease and are diagnosed correctly.}
}
\value{
A list with two components: 
  \describe{ 
   \item{Adjusted odds ratio}{Adjusted odds ratio after accounting for misclassification} 
   \item{Estimated odds ratio}{Estimated odds ratio based on (possibly) misclassified data} 
  }
}
\description{
Calculate potential impact of misclassification on an odds ratio estimate for
a single 2 x 2 table.
}
\examples{
## Example 4.2
adjusted.odds.ratio(data = breast, sensitivity = 0.9, specificity = 0.99) 
}
\references{
Newman (2001), pages 72-75, 99.
}

