% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breast.survival}
\alias{breast.survival}
\title{Breast cancer survival data (data frame)}
\format{data frame with 199 observations and 4 variables:
\describe{
 \item{time}{survival time in months}
 \item{status}{censoring variable; 0 = censored; 1 = died of breast cancer}
 \item{stage}{stage of disease: I, II, III}
 \item{receptor.level}{amount of estrogen receptor that is present in breast tissue: low, high}
}}
\source{
Newman (2001)
}
\usage{
breast.survival
}
\description{
A random sample of 199 female breast cancer patients who registered with the Northern Alberta Breast 
Cancer Registry in 1985. Entry into the cohort was restricted to women with either stage I, II, or III 
disease. This version of the breast cancer data includes the seven subjects who died of a cause other than
breast cancer.
Data are provided in different subsets and formats throughout Newman: 
Table 4.5(a), page 99 (\code{\link{breast}}); 
Table 5.3, page 126 (\code{\link{breast.receptor}}); 
Table 5.10, page 140 (\code{\link{breast.stage}}); 
Table 9.1, page 175 (\code{\link{breast.survival}})
}
\keyword{datasets}

