% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox-oakes.R
\name{cox.oakes}
\alias{cox.oakes}
\title{Cox-Oakes Test of Exponentiality}
\usage{
cox.oakes(time, status)
}
\arguments{
\item{time}{a numeric vector of survival times.}

\item{status}{a numeric vector of censoring indicators, with 0 = censored and
1 = dead.}
}
\value{
A list with class \code{"htest"} containing the following 
  components: 
  \describe{ 
   \item{statistic}{The Cox-Oakes test statistic.} 
   \item{p.value}{The p-value of the test.} 
   \item{estimate}{An estimate of the hazard rate.} 
   \item{method}{A character string indicating the method employed.} 
   \item{data.name}{A character string giving the name of the data.} 
  }
}
\description{
Performs a Cox-Oakes test of exponentiality of the null that the shape 
parameter of a Weibull distribution equals 1.
}
\details{
Under the null the Weibull distribution simplifies to an exponential 
distribution, implying a constant hazard function. Large values of the test 
statistic provide evidence against the exponential assumption of a constant 
hazard function. According to Newman (2001): "The correct interpretation is 
as follows: Given that we have decided to fit the data using using a Weibull 
model, not rejecting the null means there is no reason not to choose the 
exponential model (which is a type of Weibull model)." (p. 197)
}
\examples{
## Example 10.1
cox.oakes(time = breast.survival$time, status = breast.survival$status)
## Provides moderate evidence that exponential assumption may not be satisfied.

## A graphical assessment can also be performed by plotting the estimated 
## exponential survival curve and the Kaplan-Meier curve and deciding subjectively
## whether the latter appears to be exponential in appearance.

## Fig 10.2(a), p. 198
require(flexsurv)
fit.exp <- flexsurvreg(formula = Surv(time, status) ~ 1, data = breast.survival, dist="exp")
plot(fit.exp, ci=FALSE)

}
\references{
Newman (2001), page 197.
}

