% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{estrogen2}
\alias{estrogen2}
\title{Estrogen data (2 x 5 matrix)}
\format{
2 x 5 matrix
\describe{
\item{Case}{exposed, unexposed}
\item{Number of exposed controls}{0,1,2,3,4}
}
}
\source{
Newman (2001), Table 11.14; Mack, et al. (1976). Estrogens and endometrial cancer in a retirement
community. \emph{NEJM} 294, 1262-1267.
}
\usage{
estrogen2
}
\description{
Data from a (1:4) matched-pairs case-control study investigating estrogen use as a risk factor
for endometrial cancer.
}
\keyword{datasets}
