% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact-rate-test.R
\name{exact.rate.test}
\alias{exact.rate.test}
\title{Exact hazard rate test - single sample}
\usage{
exact.rate.test(time, status, null = 1, conf.level = 0.95, upper = 100)
}
\arguments{
\item{time}{a numeric vector of survival times.}

\item{status}{a numeric vector of censoring indicators, with 0 = censored and
1 = dead.}

\item{null}{the null hazard rate. Default is 1.}

\item{conf.level}{confidence level of the returned confidence interval. Must be a single number between 0 and 1.}

\item{upper}{upper end point of the interval to be searched for one dimensional 
root (zero) finding (passed to \code{\link[stats]{uniroot}})}
}
\value{
A list with class \code{"htest"} containing the following 
  components: 
  \describe{ 
   \item{p.value}{The p-value of the test.} 
   \item{estimate}{An estimate of the hazard rate.} 
   \item{null.value}{The null hazard rate.} 
   \item{conf.int}{A confidence interval for the hazard rate.} 
   \item{alternative}{A character string describing the alternative hypothesis. Currently only "two.sided".} 
   \item{method}{A character string indicating the method employed.} 
   \item{data.name}{A character string giving the name of the data.} 
  }
}
\description{
Performs an exact test of the null that a single sample hazard rate is equal to some value.
}
\examples{
## Example 10.3
exact.rate.test(time = 10, status = 2, null = 0.4)

## With Breast Cancer Survival data
with(breast.survival, exact.rate.test(time = time, status = status))
with(breast.survival, exact.rate.test(time = time, status = status, null = 0.01))
}
\references{
Newman (2001), page 203-4.
}
