% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOC_BME_bydate.R
\name{getOC_BME_bydate}
\alias{getOC_BME_bydate}
\title{Get operating characteristics via simulations for an biomarker enrichment design}
\usage{
getOC_BME_bydate(
  nsim = 1000,
  nF = 600,
  nS_additional = 100,
  prop_S = 0.5,
  duration = 20,
  duration_additional = 3,
  cutTime = 20,
  targetEvents = c(210, 280, 420),
  HR.Sc.threshold = 0.9,
  hazard_S = c(0.1, 0.1),
  hazard_Sc = c(0.12, 0.12),
  dropout_S = c(0, 0),
  dropout_Sc = c(0, 0),
  w = 1,
  ratio = 1,
  orr_S = c(0.2, 0.3),
  orr_Sc = c(0.2, 0.4),
  rho_S = 0.7,
  rho_Sc = 0.7,
  alpha1 = 0.0125,
  alpha2 = 0.0125,
  selection = "survival",
  orr_thres = 0.1
)
}
\arguments{
\item{nsim}{number of replicates}

\item{nF}{total number of subjects for full population.}

\item{nS_additional}{number of subjects for subgroup (BM+)}

\item{prop_S}{proportion for sub-population group S.}

\item{duration}{enrollment duration in months for full population}

\item{duration_additional}{if expand, enrollment duration in months for additional patients in BM+}

\item{targetEvents}{A numeric vector of length 3, specifying the number of events for:
final analysis in S without expansion, final analysis in S with expansion, and  final analysis in F.}

\item{HR.Sc.threshold}{Hazard ratio threshold for futility.}

\item{hazard_S}{hazard rates (h_control, h_treatment) for S}

\item{hazard_Sc}{hazard rates (h_control, h_treatment) for Sc}

\item{dropout_S}{dropout hazard rates (h_control, h_treatment) for S}

\item{dropout_Sc}{dropout hazard rates (h_control, h_treatment) for Sc}

\item{w}{weight parameter in cumulative enrollment pattern.}

\item{ratio}{randomization ratio r:1, where r refers to treatment arm; for equal randomization, r=1.}

\item{alpha1}{Significance level for subgroup S}

\item{alpha2}{Significance level for full population F}

\item{orr_thres}{}
}
\value{
It returns a list.
}
\description{
Get operating characteristics via simulations for an biomarker enrichment design
}
\examples{
getOC_BME(seed = 2025, nsim = 10, nF = 600, nS_additional = 100,  prop_S = 0.5,duration = 20)
}
