% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOC_BME_GS_bydate.R
\name{getOC_BME_GS_bydate}
\alias{getOC_BME_GS_bydate}
\title{Get operating characteristics via simulations for an biomarker enrichment design with group sequential method}
\usage{
getOC_BME_GS_bydate(
  seed = 2025,
  nsim = 1000,
  nF = 700,
  nS_additional = 150,
  prop_S = 0.5,
  duration = 23,
  duration_additional = 5,
  cutTime = 20,
  HR.Sc.threshold = 0.9,
  targetEvents.S.noexpand = c(150, 210),
  targetEvents.S.expand = c(200, 280),
  targetEvents.F = c(300, 420),
  design_S_noexpand,
  design_S_expand,
  design_F,
  hazard_S = c(0.1, 0.1),
  hazard_Sc = c(0.12, 0.12),
  dropout_S = c(0, 0),
  dropout_Sc = c(0, 0),
  w = 1,
  ratio = 1,
  orr_S = c(0.2, 0.3),
  orr_Sc = c(0.2, 0.4),
  rho_S = 0.7,
  rho_Sc = 0.7,
  alpha1 = 0.0125,
  alpha2 = 0.0125,
  selection = "survival",
  orr_thres = 0.1
)
}
\arguments{
\item{seed}{random seed for reproducibility}

\item{nsim}{number of replicates}

\item{nF}{total number of subjects for full population.}

\item{nS_additional}{number of subjects for subgroup (BM+)}

\item{prop_S}{proportion for sub-population group S.}

\item{duration}{enrollment duration in months for full population}

\item{duration_additional}{if expand, enrollment duration in months for additional patients in BM+}

\item{HR.Sc.threshold}{Hazard ratio threshold for futility.}

\item{targetEvents.S.noexpand}{The target number of events in S without expansion}

\item{targetEvents.S.expand}{The target number of events in S with expansion}

\item{targetEvents.F}{The target number of events in F}

\item{hazard_S}{hazard rates (h_control, h_treatment) for S}

\item{hazard_Sc}{hazard rates (h_control, h_treatment) for Sc}

\item{dropout_S}{dropout hazard rates (h_control, h_treatment) for S}

\item{dropout_Sc}{dropout hazard rates (h_control, h_treatment) for Sc}

\item{w}{weight parameter in cumulative enrollment pattern.}

\item{ratio}{randomization ratio r:1, where r refers to treatment arm; for equal randomization, r=1.}

\item{orr_S}{Objective response (binary: 1 = response, 0 = non-response) for S}

\item{orr_Sc}{Objective response (binary: 1 = response, 0 = non-response) for Sc}

\item{rho_S}{Correlation between ORR and time to event endpoint for S}

\item{rho_Sc}{Correlation between ORR and time to event endpoint for Sc}

\item{alpha1}{Significance level for subgroup S}

\item{alpha2}{Significance level for full population F}

\item{selection}{"survival" or "orr:}

\item{targetEvents.Sc}{The target number of events in Sc; length of 1,
futility or population selection will be given at IA without efficacy testing.}

\item{orr_number}{}
}
\value{
It returns a list.
}
\description{
Get operating characteristics via simulations for an biomarker enrichment design with group sequential method
}
\examples{
getOC_BME(seed = 2025, nsim = 10, nF = 600, nS_additional = 100,  prop_S = 0.5,duration = 20)
}
