% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOC_data.R
\name{getOC_data}
\alias{getOC_data}
\title{Title}
\usage{
getOC_data(
  seed = 2025,
  nsim = 1000,
  nF = 600,
  prop_S = 0.5,
  duration = 20,
  hazard_S = c(0.1, 0.1),
  hazard_Sc = c(0.12, 0.12),
  dropout_S = c(0, 0),
  dropout_Sc = c(0, 0),
  w = 1,
  ratio = 1,
  orr_S = c(0.2, 0.3),
  orr_Sc = c(0.2, 0.4),
  rho_S = 0.7,
  rho_Sc = 0.7,
  enroll_month = 19
)
}
\arguments{
\item{enroll_month}{}
}
\description{
Title
}
