% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZstats_FA.R
\name{getZstats_FA}
\alias{getZstats_FA}
\title{Get Z-statistics for a given trial at final analysis}
\usage{
getZstats_FA(dat_initial, dat_additional, targetEvents, expand)
}
\arguments{
\item{dat_initial}{Initial time-to-event dataset returned from \link{simu_enrich_trial}.}

\item{dat_additional}{Additional time-to-event dataset returned from \link{simu_enrich_trial}.}

\item{targetEvents}{A numeric vector of length 3, specifying the number of events for: FA_S_noexpand, FA_S_expand, FA_F.}

\item{expand}{0=No, 1=Yes}
}
\value{
It returns a list of test statistics used for later adjustments.
}
\description{
Get Z-statistics for a given trial at final analysis
}
\examples{
d1 <- simu_enrich_trial(n = 600, prop_S = 0.5, duration = 20)
d2 <- simu_enrich_trial(n = 100, prop_S = 0.5, duration = 20)
getZstats_FA(d1, d2, targetEvents = c(210, 280, 420), expand=0)
}
