% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZstats_GS.R
\name{getZstats_GS}
\alias{getZstats_GS}
\title{Get Z-statistics for Group Sequential method}
\usage{
getZstats_GS(dat, targetEvents)
}
\arguments{
\item{dat}{A time-to-event dataset returned from \link{simu_enrich_trial}.}

\item{targetEvents}{The target number of events}
}
\value{
It returns a list of Z statistics used for later adjustments and observed events number.
}
\description{
Get Z-statistics for Group Sequential method
}
\examples{
d <- simu_enrich_trial(n = 200, prop_S = 0.5, duration = 10)
targetEvents <- c(200, 300)
getZstats_GS(d, targetEvents)
}
