% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZstats_IA.R
\name{getZstats_IA}
\alias{getZstats_IA}
\title{Get Z-statistics for a given trial at interim analysis.}
\usage{
getZstats_IA(dat, targetEvents.Sc)
}
\arguments{
\item{dat}{A time-to-event dataset returned from \link{simu_enrich_trial}.}

\item{targetEvents.Sc}{The target number of events in Sc; length of 1,
futility or population selection will be given at IA without efficacy testing.}
}
\value{
It returns a list of test statistics used for later adjustments.
}
\description{
Get Z-statistics for a given trial at interim analysis.
}
\examples{
d <- simu_enrich_trial(n = 600, prop_S = 0.5, duration = 20)
getZstats_IA(d, targetEvents.Sc = 105)
}
