% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZstats_IA_bydate.R
\name{getZstats_IA_bydate}
\alias{getZstats_IA_bydate}
\title{Get Z-statistics for a given trial at interim analysis.}
\usage{
getZstats_IA_bydate(dat, cutTime)
}
\arguments{
\item{dat}{A time-to-event dataset returned from \link{simu_enrich_trial}.}
}
\value{
It returns a list of test statistics used for later adjustments.
}
\description{
Get Z-statistics for a given trial at interim analysis.
}
\examples{
d <- simu_enrich_trial(n = 600, prop_S = 0.5, duration = 20)
getZstats_IA(d, targetEvents.Sc = 105)
}
