% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZtest_GS.R
\name{getZtest_GS}
\alias{getZtest_GS}
\title{Get test statistics for Group Sequential method}
\usage{
getZtest_GS(Zstats, design)
}
\arguments{
\item{dat_S}{Dataset for subgroup S (BM+), with columns: surtime, event, trt}

\item{design_S}{Group sequential design object from rpact::getDesignGroupSequential()}

\item{n_target}{Total number of events planned for final analysis}
}
\value{
A list with Z-statistics, observed events, and rejection decision
}
\description{
Get test statistics for Group Sequential method
}
\examples{
d <- simu_enrich_trial(n = 200, prop_S = 0.5, duration = 10)
design <- getDesignGroupSequential(kMax = 2, alpha = 0.0125, informationRates = c(0.5, 1))
Zstats <- getZstats_GS(d, targetEvents = c(100,200))
getZtest_GS(Zstats, design)
}
