% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZtests_BME.R
\name{getZtests_BME}
\alias{getZtests_BME}
\title{Get adaptive enrichment design test statistics for a given trial.}
\usage{
getZtests_BME(zstats, alpha1 = 0.0125, alpha2 = 0.0125)
}
\arguments{
\item{alpha1}{Significance level for subgroup S}

\item{alpha2}{Significance level for full population F}

\item{zstatsAn}{object returned by getZstats_FA()}
}
\value{
It returns a list of test statistics used for later adjustments.
}
\description{
Get adaptive enrichment design test statistics for a given trial.
}
