% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio-test.R
\name{hazard.ratio.test}
\alias{hazard.ratio.test}
\title{Hazard Ratio Test of association for unstratified survival data}
\usage{
hazard.ratio.test(time, status, exposure, Wald = TRUE, conf.level = 0.95,
  exact = FALSE)
}
\arguments{
\item{time}{a numeric vector of survival times.}

\item{status}{a numeric vector of censoring indicators, with 0 = censored and
1 = dead.}

\item{exposure}{a factor vector with two levels indicating exposure. The
first level is assumed to be the exposed condition.}

\item{Wald}{a logical indicating whether to use the Wald or Likelihood Ratio 
Test. Default is TRUE.}

\item{conf.level}{conf.level confidence level of the returned confidence 
interval. Must be a single number between 0 and 1. Default is 0.95.}

\item{exact}{a logical indicating whether the asymptotic (unconditional) test
or the exact test should be computed. Default is FALSE. The \code{"Wald"}
argument is ignored when set to TRUE.}

\item{rev}{a logical indicating to whether to reverse the order of the factor
levels when estimating the hazard ratio.}
}
\value{
A list with class \code{"htest"} containing the following 
  components: 
  \describe{ 
   \item{statistic}{The chi-square test statistic. Only returned if \code{"exact = FALSE"}.} 
   \item{p.value}{The p-value of the test.} 
   \item{estimate}{Hazard ratio estimate.} 
   \item{null.value}{The null hazard ratio, which is currently set to 1.} 
   \item{alternative}{A character string describing the alternative hypothesis. Currently only "two.sided".} 
   \item{method}{A character string indicating the method employed.} 
   \item{data.name}{A character string giving the name of the data.} 
  }
}
\description{
Performs a test of the null that the hazard ratio between two groups is 1.
}
\examples{
## Example 10.8
with(breast.survival, hazard.ratio.test(time, status, receptor.level))
with(breast.survival, hazard.ratio.test(time, status, receptor.level, Wald = FALSE))
## Example 10.9
## Receptor Level-Breast Cancer: Stage III
dat <- subset(breast.survival, stage=="III")
hazard.ratio.test(time = dat$time, status = dat$status, exposure = dat$receptor.level)
hazard.ratio.test(time = dat$time, status = dat$status, exposure = dat$receptor.level, 
                  Wald = FALSE)
## Example 10.10
hazard.ratio.test(time = dat$time, status = dat$status, exposure = dat$receptor.level, 
                  exact = TRUE)
}
\references{
Newman (2001), page 206 - 213.
}

