% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard-ratios.R
\name{hazard.ratios}
\alias{hazard.ratios}
\title{Hazard ratios of stratified survival data}
\usage{
hazard.ratios(time, status, exposure, strata, method = "wald")
}
\arguments{
\item{time}{a numeric vector of survival times.}

\item{status}{a numeric vector of censoring indicators, with 0 = censored and
1 = dead.}

\item{exposure}{a factor vector with two levels indicating exposure. The 
first level is assumed to be the exposed condition.}

\item{strata}{a factor vector with at least two levels indicating strata.}

\item{method}{method for calculating hazard ratio and confidence interval.
Choices are "wald" (default) and "midp".}
}
\value{
A list containing a hazard ratio estimate and confidence interval for each stratum.
}
\description{
Returns hazard ratios and confidence intervals for each stratum.
}
\details{
Hazard ratios and their confidence intervals are calculated with the Wald 
method using \code{epitools::rateratio}. This can be changed to 
median-unbiased estimation by setting \code{method="midp"}.
}
\examples{
## Table 10.15 from Example 10.14
with(breast.survival, hazard.ratios(time, status, receptor.level, stage))

## using the "midp" method
with(breast.survival, hazard.ratios(time, status, receptor.level, stage, method="midp"))
}
\references{
Newman (2001), page 208.
}
