% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratio-k-test.R
\name{k.hazard.ratio.test}
\alias{k.hazard.ratio.test}
\title{Hazard Ratio Test of association for survival data with k strata}
\usage{
k.hazard.ratio.test(time, status, exposure, strata, Wald = TRUE,
  conf.level = 0.95)
}
\arguments{
\item{time}{a numeric vector of survival times.}

\item{status}{a numeric vector of censoring indicators, with 0 = censored and
1 = dead.}

\item{exposure}{a factor vector with two levels indicating exposure. The
first level is assumed to be the exposed condition.}

\item{strata}{a factor vector with at least two levels indicating strata.}

\item{Wald}{a logical indicating whether to use the Wald or Likelihood Ratio 
Test. Default is TRUE.}

\item{conf.level}{conf.level confidence level of the returned confidence 
interval. Must be a single number between 0 and 1. Default is 0.95.}
}
\value{
A list with class \code{"htest"} containing the following 
  components: 
  \describe{ 
  \item{statistic}{The chi-square test statistic.} 
  \item{p.value}{The p-value of the test.} 
  \item{estimate}{Common hazard ratio estimate.} 
  \item{null.value}{The null common hazard ratio, which is currently set to 1.} 
  \item{alternative}{A character string describing the alternative hypothesis. Currently only "two.sided".} 
  \item{method}{A character string indicating the method employed.} 
  \item{data.name}{A character string giving the name of the data.} }
}
\description{
Performs a test of the null that the hazard ratio between \emph{k} strata is 1.
}
\examples{
## Example 10.14
with(breast.survival, 
     k.hazard.ratio.test(time = time, status = status, 
                         exposure = receptor.level, strata = stage))
with(breast.survival, 
     k.hazard.ratio.test(time = time, status = status,
                         exposure = receptor.level, strata = stage, 
                         Wald=FALSE))
}
\references{
Newman (2001), page 219 - 221.
}

