% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-trend.R
\name{linear.trend.test}
\alias{linear.trend.test}
\title{Test for linear trend (in log-hazard ratios)}
\usage{
linear.trend.test(time, status, exposure, strata,
  scores = seq_along(length(levels(strata))))
}
\arguments{
\item{time}{a numeric vector of survival times.}

\item{status}{a numeric vector of censoring indicators, with 0 = censored and
1 = dead.}

\item{exposure}{a factor vector with two levels indicating exposure. The 
first level is assumed to be the exposed condition.}

\item{strata}{a factor vector with at least two levels indicating strata.}

\item{score}{group score.}
}
\value{
An object of class "htest" with title, test statistic, and p-value.
}
\description{
Performs a test for linear trend of log-hazard ratios in stratified survival data.
}
\examples{
## Example 10.14
with(breast.survival, linear.trend.test(time, status, receptor.level, stage)) 

## Using specified scores
with(breast.survival, linear.trend.test(time, status, receptor.level, stage,
                                        scores = c(1,2,6))) 
}
\references{
Newman (2001), page 221.
}

