% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt-homogeneity.R
\name{lrt.homogeneity}
\alias{lrt.homogeneity}
\title{Likelihood ratio test of homogeneity of hazard ratios (stratified survival 
data)}
\usage{
lrt.homogeneity(time, status, exposure, strata)
}
\arguments{
\item{time}{a numeric vector of survival times.}

\item{status}{a numeric vector of censoring indicators, with 0 = censored and
1 = dead.}

\item{exposure}{a factor vector with two levels indicating exposure. The 
first level is assumed to be the exposed condition.}

\item{strata}{a factor vector with at least two levels indicating strata.}
}
\value{
An object of class "htest" with title, test statistic, and p-value.
}
\description{
Performs a likelihood ratio test of homogeneity of hazard ratios for 
stratified survival data. The null hypothesis is that all stratum-specific 
hazard ratios are equal.
}
\details{
According to Newman (p. 124): "the homogeneity assumption is merely a 
convenient fiction that is adopted in order to simplify the analysis and 
interpretation of data." A suggested approach is to first examine 
stratum-specific hazard ratios and their confidence intervals to get a sense 
if there are meaningful differences across strata (see 
\code{\link{hazard.ratios}}); perform a formal test of homogeneity; and then 
"synthesize this information along with substantive knowledge of the 
relationship between exposure and disease, taking into account the aims of
the study."
}
\examples{
## Example 10.14
with(breast.survival, lrt.homogeneity(time, status, receptor.level, stage)) 
}
\references{
Newman (2001), page 221.
}
