% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh-mpcc-test.R
\name{mantelhaen.mpcc.test}
\alias{mantelhaen.mpcc.test}
\title{Mantel-Haenszel Test of association for matched-pairs case-control}
\usage{
mantelhaen.mpcc.test(data, conf.level = 0.95)
}
\arguments{
\item{data}{a matrix with two rows and at least two columns with exposed cases on the 
first row and exposed controls on the first column (see data used in examples below)}

\item{conf.level}{confidence level of the returned confidence 
interval. Must be a single number between 0 and 1. Default is 0.95.}
}
\value{
A list with class \code{"htest"} containing the following 
  components: 
  \describe{ 
   \item{statistic}{The Mantel-Haenszel chi-square test statistic.} 
   \item{parameter}{The degrees of freedom of the test statistic.}
   \item{p.value}{The p-value of the test.} 
   \item{estimate}{The Mantel-Haenszel odds ratio estimate.} 
   \item{null.value}{The null hazard ratio, which is currently set to 1.} 
   \item{alternative}{A character string describing the alternative hypothesis. Currently only "two.sided".} 
   \item{method}{A character string indicating the method employed.} 
   \item{data.name}{A character string giving the name of the data.} 
  }
}
\description{
Performs the Mantel-Haenszel Test of association for matched-pairs case-control data.
}
\examples{
## Example 11.3 - (1:1) matched case-control data
estrogen
mantelhaen.mpcc.test(estrogen)

## Example 11.4 - (1:M) matched case-control data
estrogen2
mantelhaen.mpcc.test(estrogen2)
}
\references{
Newman (2001), pages 243 - 246.
}
\seealso{
\code{\link[stats]{mcnemar.test}}
}

