% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdlt.R
\name{mdlt}
\alias{mdlt}
\title{Multiple Decrement Life Table (MDLT) for a given year}
\usage{
mdlt(group, all.deaths, cause.deaths, pop, radix = 1e+05)
}
\arguments{
\item{group}{The age group vector as integer. For example, 0, 1, 5, ... where 0 indicates birth to one year, 
1 indicates age 1 to age 5, and so forth. See data used with example below.}

\item{all.deaths}{The number of deaths from all causes for each age group as a numeric vector.}

\item{cause.deaths}{The number of deaths from a particular cause for each age group as a numeric vector.}

\item{pop}{The population for each age group as a numeric vector.}

\item{radix}{The number of individuals in the OLT birth cohort. Typically set to a large number like 100,000.}
}
\value{
a \code{data.frame} with \code{length(group)} rows and 7 columns:
\describe{
\item{group}{Age group}
\item{qjk}{crude conditional probability of dying in age group}
\item{lkxj}{number of survivors to age x}
\item{djk}{number of deaths in age group}
\item{Ljk}{number of person years in age group}
\item{Tkxj}{number of person years after age x}
\item{ekxj}{life expectancy at age x}

}
}
\description{
Construct a multiple decrement life table for a given year. A multiple decrement life table 
(MDLT) describes the mortality experience of the group of individuals in the 
OLT cohort who are "due to die" of a particular cause of death. The MDLT is
an example of a competing risks model.
}
\examples{
## Example 13.3
## create multiple decrement life table for neoplasms
with(males, mdlt(group = age.group, all.deaths = deaths.all, 
                 cause.deaths = neoplasm.deaths, pop = pop))

mdlt.out <- with(males,
                 mdlt(group = age.group, all.deaths = deaths.all,
                 cause.deaths = neoplasm.deaths, pop = pop))
                 
## 27.17\% of males born in Canada in 1991 predicted to die of neoplasm
mdlt.out[1,"Ljk"] / 1e5 * 100

## For an individual due to die of neoplasm, the life expectancy 
## at birth is 73.27 years
mdlt.out[1,"ekxj"]
}
\references{
Newman (2001), pages 270-273.
}
\seealso{
\code{\link{olt}} and \code{\link{cdlt}}
}
