% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olt.R
\name{olt}
\alias{olt}
\title{Ordinary Life Table (OLT) for a given year}
\usage{
olt(group, deaths, pop, radix = 1e+05, phi = 1, linear = TRUE)
}
\arguments{
\item{group}{The age group vector as integer. For example, 0, 1, 5, ... where 0 indicates birth to one year, 
1 indicates age 1 to age 5, and so forth. See data used with example below.}

\item{deaths}{The number of deaths from all causes for each age group as a numeric vector.}

\item{pop}{The population for each age group as a numeric vector.}

\item{radix}{The number of individuals in the OLT birth cohort. Typically set to a large number like 100,000.}

\item{phi}{The multiplier for sensitivity analysis. For example, set to 0.9 to examine mortality 
implications for a declining death rate.}

\item{linear}{logical indicating which functional form to use to calculate number of survivors to age x (lxj).
Default is TRUE, which means use linear form. Otherwise use exponential.}
}
\value{
a \code{data.frame} with \code{length(group)} rows and 8 columns:
\describe{
\item{group}{Age group}
\item{qj}{conditional probability of dying }
\item{pj}{conditional probability of surviving}
\item{lxj}{number of survivors to age x}
\item{dj}{number of deaths in age group}
\item{Lj}{number of person years in age group}
\item{Txj}{Number of person years after age x}
\item{exj}{Life expectancy at age x}
}
}
\description{
Construct an ordinary life table for a given year. The ordinary life table (OLT) provides a method of analyzing
mortality for all causes of death combined.
}
\examples{
## Example 13.1
with(males, olt(group = age.group, deaths = deaths.all, pop = pop))
## predicted life expectancy birth is 73.34

## with sensitivity analysis for declinind death rate:
with(males, olt(group = age.group, deaths = deaths.all, pop = pop, phi = 0.9))
## predicted life expectancy birth is 75.65
}
\references{
Newman (2001), pages 264-269.
}
\seealso{
\code{\link{cdlt}} and \code{\link{mdlt}}
}
