% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ovarian.cancer}
\alias{ovarian.cancer}
\title{Ovarian cancer data}
\format{data frame with 35 obervations and 3 variables:
\describe{
 \item{grade}{indicator of the malignant potential of the tumor: High, Low}
 \item{time}{survival time measured in days}
 \item{status}{censoring variable; 0 = censored; 1 = died of ovarian cancer}
}}
\source{
Newman (2001), Table 9.6; Fleming, et al. Modified Kolmogorov-Smirnov test procedures with application to arbitrarily 
right-censored data. \emph{Biometrics} 36, 607-625.
}
\usage{
ovarian.cancer
}
\description{
Data from a cohort study of women with stage II or IIIA ovarian cancer, where the endpoint is progression of
disease.
}
\keyword{datasets}

