% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power-closed-cohort.R
\name{power.closed.cohort}
\alias{power.closed.cohort}
\title{Power calculations for unstratified closed cohort studies}
\usage{
power.closed.cohort(n = NULL, delta, p2, rho = 1, sig.level = 0.05,
  power = NULL, type = c("RD", "RR", "OR"), alternative = c("two.sided",
  "one.sided"), tol = .Machine$double.eps^0.25)
}
\arguments{
\item{n}{number of observations}

\item{delta}{The true risk difference (RD), risk ratio (RR) or odds ratio (OR), depending on \code{type} selected.}

\item{p2}{probability that someone without a history of exposure will develop the disease}

\item{rho}{The ratio of unexposed to exposed subjects}

\item{sig.level}{significance level (Type I error probability)}

\item{power}{power of test (1 minus Type II error probability)}

\item{type}{type of measurement: \code{RD} = "Risk Difference", \code{RR} = "Risk Ratio", and \code{OR} = "Odds Ratio"}

\item{alternative}{one- or two-sided test. Can be abbreviated.}

\item{tol}{numerical tolerance used in root finding, the default providing (at least) four significant digits. 
Root finding refers to \code{\link[stats]{uniroot}}, which is used to find sample size given power.}
}
\value{
Object of class "power.htest", a list of the arguments (including the computed one) augmented 
with method and note elements.
}
\description{
Compute the power of risk difference, risk ratio and odds ratio methods for closed cohort studies, 
or determine parameters to obtain a target power.
}
\details{
Exactly one of the parameters n and power must be passed as NULL, and that parameter is determined 
from the others.
}
\examples{
## Example 14.2 (row 1 of Table 14.2)
power.closed.cohort(delta = 0.01, p2 = 0.05, rho = 1, type = "RD", power = 0.80)

## Table 14.2
sapply(c(0.01, 0.05, 0.10, 0.20, 0.30), 
       function(x)power.closed.cohort(delta = x, p2 = 0.05, rho = 1, power = 0.8)$r1)
       
## Table 14.3
sapply(lapply(c(1:5,10,20), 
       function(x)power.closed.cohort(delta = 0.05, rho = x, power = 0.8, p2 = 0.05)),
function(x)x[1:2])

## Example 14.5 (row 2 of Table 14.4)
power.closed.cohort(delta = 3, p2 = 0.05, rho = 1, power = 0.8, type = "OR")

## Table 14.4
sapply(c(2:5,10), function(x)power.closed.cohort(delta = x, p2 = 0.05, 
                                                 rho = 1, power = 0.8, type = "OR")$r1)

## Example 14.7
power.closed.cohort(n = 100, delta = 2, p2 = 0.10, rho = 2, type = "OR")

## Example 14.8
power.closed.cohort(rho = 6.73, p2 = 0.048, delta = 3, power = 0.8, type = "OR")
}
\references{
Newman (2001), pages 283 - 285
}
\seealso{
\code{\link[stats]{power.prop.test}} which can perform Risk Difference power calculations assuming \code{rho = 1}
}

