% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power-hr-test.R
\name{power.hr.test}
\alias{power.hr.test}
\title{Power calculation for Hazard Ratio Test}
\usage{
power.hr.test(n = NULL, hr, p1, pi2, sig.level = 0.05, power = NULL,
  alternative = c("two.sided", "one.sided"), tol = .Machine$double.eps^0.25)
}
\arguments{
\item{n}{number of subjects needed for study}

\item{hr}{Hazard Ratio (ie, the "ratio worth detecting")}

\item{p1}{proportion of cohort with history of exposure}

\item{pi2}{probability a member of unexposed cohort will die during follow-up}

\item{sig.level}{significance level (Type I error probability)}

\item{power}{power of test (1 minus Type II error probability)}

\item{alternative}{one- or two-sided test. Can be abbreviated.}

\item{tol}{numerical tolerance used in root finding, the default providing (at least) 
four significant digits. Root finding refers to \code{\link[stats]{uniroot}}, which is used to find sample size given power.}
}
\value{
Object of class "power.htest", a list of the arguments (including the computed one) 
augmented with method and note elements.
}
\description{
Compute the power of the Hazard Ratio Test, or determine parameters to obtain a target power.
}
\details{
Exactly one of the parameters n and power must be passed as NULL, and 
that parameter is determined from the others.
}
\examples{
## Example 14.4
power.hr.test(hr = 2, p1 = 0.251, pi2 = 0.174, power = 0.8)
power.hr.test(n = 414, hr = 2, p1 = 0.251, pi2 = 0.174)
}
\references{
Newman (2001), pages 286 - 287
}
\seealso{
\code{\link{hazard.ratio.test}}
}

