% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power-mpcc.R
\name{power.mpcc}
\alias{power.mpcc}
\title{Power Calculation for Matched-Pairs Case-Control Odds Ratio Test}
\usage{
power.mpcc(n = NULL, OR, p2, v, sig.level = 0.05, power = NULL, M = 1,
  alternative = c("two.sided", "one.sided"), tol = .Machine$double.eps^0.25)
}
\arguments{
\item{n}{number of matched pairs needed for study}

\item{OR}{Odds Ratio (ie, the "ratio worth detecting")}

\item{p2}{probability a control has a history of exposure}

\item{v}{measure of concordance (of exposure) in the case and control samples. 
\code{v} = 1 if and only if the matching variables are not associated with exposure in the case and control samples.}

\item{sig.level}{significance level (Type I error probability)}

\item{power}{power of test (1 minus Type II error probability)}

\item{M}{number of controls per case (default = 1)}

\item{alternative}{one- or two-sided test. Can be abbreviated.}

\item{tol}{numerical tolerance used in root finding, the default providing (at least) four significant digits. 
Root finding refers to \code{\link[stats]{uniroot}}, which is used to find sample size given power.}
}
\value{
Object of class "power.htest", a list of the arguments (including the computed one) augmented with 
method and note elements.
}
\description{
Compute the power of the Matched-Pairs Case-Control Odds Ratio Test, or determine parameters to obtain a target power.
}
\details{
Exactly one of the parameters n and power must be passed as NULL, and that parameter is determined from the others.

When pre-study information on concordance is limited, Fleiss and Levin (1988) recommend using a relatively 
large value of \code{v}, such as 2.5.

Sample size estimates for M >=1 can be overestimated when the probability is small that a sample control 
has a history of exposure.
}
\examples{
## Example 14.6
power.mpcc(power = 0.80, OR = 3, p2 = 0.05, v = 4.82)
}
\references{
Newman (2001), pages 288 - 291

Dupont, WD. (1988) Power calculations for matched case-control studies. \emph{Biometrics} 44, 1157-1168.

Fleiss, JL and Levin B. (1988) Sample size determination in studies with matched pairs. \emph{Journal of Clinical Epidemiology} 41, 727-730.
}

