% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power-or-ucc.R
\name{power.or.ucc}
\alias{power.or.ucc}
\title{Power calculations for Odds Ratio test in a unmatched case control study}
\usage{
power.or.ucc(n = NULL, power = NULL, rho = 1, p2, OR, sig.level = 0.05,
  tol = .Machine$double.eps^0.25)
}
\arguments{
\item{n}{number of cases}

\item{power}{power of test (1 minus Type II error probability)}

\item{rho}{ratio of controls to cases}

\item{p2}{probability that a control has a history of exposure}

\item{OR}{odds ratio}

\item{sig.level}{significance level (Type I error probability)}

\item{tol}{numerical tolerance used in root finding, the default providing (at least) four significant digits.}
}
\value{
Object of class "power.htest", a list of the arguments (including the computed one) 
augmented with method and note elements.
}
\description{
Compute the power of an odds ratio test or determine the sample size to obtain a target power.
}
\details{
Exactly one of the parameters \code{n} and \code{power} must be passed as NULL, and that
parameter is determined from the others.

Currently, only the two-sided alternative is calculated.
}
\examples{
## Example 14.5 (Table 14.4)
sapply(c(2,3,4,5,10), function(x) power.icc(power = 0.8, rho = 1, p2 = 0.05, OR = x)$n)

## Example 14.7
power.icc(n = 100, rho = 2, OR = 2, p2 = 0.10)

## Example 14.8
power.icc(power = 0.8, rho = 6.73, OR = 3, p2 = 0.048)
}
\references{
Newman (2001) pages 287, 293.
}

