% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power-smr-test.R
\name{power.smr.test}
\alias{power.smr.test}
\title{Power calculation for Standardized Mortality Ratio (SMR) Test}
\usage{
power.smr.test(
  n = NULL,
  smr,
  r,
  sig.level = 0.05,
  power = NULL,
  alternative = c("two.sided", "one.sided"),
  tol = .Machine$double.eps^0.25
)
}
\arguments{
\item{n}{n is amount of person-time needed for study}

\item{smr}{Stanardized Mortality Ratio (ie, the "ratio worth detecting")}

\item{r}{death rate in the standard population}

\item{sig.level}{significance level (Type I error probability)}

\item{power}{power of test (1 minus Type II error probability)}

\item{alternative}{one- or two-sided test. Can be abbreviated.}

\item{tol}{numerical tolerance used in root finding, the default providing (at least) four significant digits. 
Root finding refers to \code{\link[stats]{uniroot}}, which is used to find sample size given power.}
}
\value{
Object of class "power.htest", a list of the arguments (including the computed one) augmented 
with method and note elements.
}
\description{
Compute the power of the Standardized Mortality Ratio (SMR) Test, or determine parameters to 
obtain a target power.
}
\details{
Exactly one of the parameters n and power must be passed as NULL, and that parameter is determined 
from the others.
}
\examples{
## Example 14.3
power.smr.test(smr = 1.5, r = 7283/957247, power = 0.8)

}
\references{
Newman (2001), pages 285 - 286.
}
\seealso{
\code{\link{smr.test}}
}
