% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate-test.R
\name{rate.test}
\alias{rate.test}
\title{Asymptotic hazard rate test - single sample}
\usage{
rate.test(time, status, null = 1, conf.level = 0.95, explicit = TRUE)
}
\arguments{
\item{time}{a numeric vector of survival times.}

\item{status}{a numeric vector of censoring indicators, with 0 = censored and
1 = dead.}

\item{null}{the null hazard rate. Default is 1.}

\item{conf.level}{confidence level of the returned confidence interval. Must be a single number between 0 and 1.}

\item{explicit}{a logical indicating whether to calculate an explicit or implicit confidence interval. Default is TRUE.}
}
\value{
A list with class \code{"htest"} containing the following 
  components: 
  \describe{ 
   \item{statistic}{The chi-square test statistic.} 
   \item{p.value}{The p-value of the test.} 
   \item{estimate}{An estimate of the hazard rate.} 
   \item{null.value}{The null hazard rate.} 
   \item{conf.int}{A confidence interval for the hazard rate.} 
   \item{alternative}{A character string describing the alternative hypothesis. Currently only "two.sided".} 
   \item{method}{A character string indicating the method employed.} 
   \item{data.name}{A character string giving the name of the data.} 
  }
}
\description{
Performs an asymtotic test of the null that a single sample hazard rate is equal to some value.
}
\examples{
## Examples 10.5 and 10.6
rate.test(time = 10, status = 2, null = 0.4)
rate.test(time = 25, status = 5, null = 0.4)
rate.test(time = 50, status = 10, null = 0.4)

rate.test(time = 10, status = 2, null = 0.4, explicit = FALSE)
rate.test(time = 25, status = 5, null = 0.4, explicit = FALSE)
rate.test(time = 50, status = 10, null = 0.4, explicit = FALSE)

## Example 10.7
with(breast.survival, rate.test(time = time, status = status, explicit = FALSE))

}
\references{
Newman (2001), page 205.
}

