% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk-ratio-test.R
\name{risk.ratio.test}
\alias{risk.ratio.test}
\title{Risk Ratio Test for single 2 x 2 table}
\usage{
risk.ratio.test(data, conf.level = 0.95, Wald = TRUE)
}
\arguments{
\item{data}{a 2 x 2 matrix.}

\item{conf.level}{confidence level of the returned confidence 
interval. Must be a single number between 0 and 1. Default is 0.95.}

\item{Wald}{a logical indicating whether to use the Wald or Likelihood Ratio 
Test. Default is TRUE.}
}
\value{
A list with class \code{"htest"} containing the following 
  components: 
  \describe{ 
  \item{statistic}{The chi-square test statistic.} 
  \item{parameter}{The degrees of freedom of the approximate chi-squared distribution of the test statistic.}
  \item{p.value}{The p-value of the test.} 
  \item{estimate}{The unconditional maximum likelihood estimate of the risk ratio.} 
  \item{null.value}{The null risk ratio, which is currently set to 1.} 
  \item{alternative}{A character string describing the alternative hypothesis. Currently only "two.sided".} 
  \item{method}{A character string indicating the method employed.} 
  \item{data.name}{A character string giving the name of the data.} }
}
\description{
Performs a test of the null that the risk ratio between two groups in a 2 x 2 table is 1.
}
\examples{
## Example 6.1
risk.ratio.test(breast)
risk.ratio.test(breast, Wald = FALSE)
}
\references{
Newman (2001), pages 143-144.
}
\seealso{
\code{\link[epitools]{riskratio}} in the epitools package.
}
