% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smr-test.R
\name{smr.test}
\alias{smr.test}
\title{Standardized Mortality Ratio (SMR) Test}
\usage{
smr.test(count, pop, stdcount, stdpop, conf.level = 0.95)
}
\arguments{
\item{count}{vector of age-specific cohort deaths}

\item{pop}{vector of age-specific cohort population or person-years}

\item{stdcount}{vector of age-specific standard population deaths}

\item{stdpop}{vector of age-specific standard population}

\item{conf.level}{confidence level as a number between 0 and 1. Default is 0.95.}
}
\value{
A list with class \code{"htest"} containing the following 
  components: 
  \describe{ 
   \item{p.value}{The p-value of the test.} 
   \item{estimate}{An estimate of the SMR ratio.} 
   \item{null.value}{The null SMR ratio. Currently set to 1.} 
   \item{conf.int}{A confidence interval for the SMR ratio.} 
   \item{alternative}{A character string describing the alternative hypothesis. Currently only "two.sided".} 
   \item{method}{A character string indicating the method employed.} 
   \item{data.name}{Character strings giving the name of the data.} 
  }
}
\description{
Performs a chi-sqaure test of the null that there is no mortality difference 
between the cohort and standard population.
}
\details{
When the expected number of deaths is less than 5, \code{\link{exact.rate.test}} 
is used to calculate the confidence interval and p-value.
}
\examples{
## Example 12.3

with(schizophrenia, 
     smr.test(count = cohort.deaths,
              pop = cohort.py, 
              stdcount = alberta.deaths, 
              stdpop = alberta.pop))
                        
## Example 12.3 continued,
## Exact test (expected number of deaths < 5)

with(subset(schizophrenia,age.group=="10-19"), 
     smr.test(count = cohort.deaths,
              pop = cohort.py, 
              stdcount = alberta.deaths, 
              stdpop = alberta.pop))

}
\references{
Newman (2001), page 255-258.
}
\seealso{
\code{\link[epitools]{ageadjust.indirect}} in the \code{epitools} package.
}

