% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std-rate-ratio.R
\name{std.rate.ratio}
\alias{std.rate.ratio}
\title{Standardized Rate Ratio}
\usage{
std.rate.ratio(count, pop, stdcount, stdpop, conf.level = 0.95)
}
\arguments{
\item{count}{vector of age-specific cohort deaths}

\item{pop}{vector of age-specific cohort population or person-years}

\item{stdcount}{vector of age-specific standard population deaths}

\item{stdpop}{vector of age-specific standard population}

\item{conf.level}{confidence level as a number between 0 and 1. Default is 0.95.}
}
\value{
a list with the following components:
\describe{
\item{Standardized rate ratio}{The ratio of the standardized death rates}
\item{CI}{The confidence interval for the standardized rate ratio}
\item{Crude rate ratio}{The ratio of the crude cohort death rate to the standard population death rate}
\item{directly standardized death rate}{The weighted average of the cohort age-specific death rates}
}
}
\description{
Calculate the ratio of standardized death rates and its confidence interval.
}
\examples{
with(schizophrenia, 
     std.rate.ratio(count = cohort.deaths, 
                    pop = cohort.py,
                    stdcount = alberta.deaths, 
                    stdpop = alberta.pop))

}
\references{
Newman (2001), pages 251-255.
}
\seealso{
\code{\link[epitools]{ageadjust.direct}} in the \code{epitools} package.
}

