% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{breast.stage}
\alias{breast.stage}
\title{Breast cancer data, stratified by disease stage (2 x 2 x 3 array)}
\format{\code{breast} 2 x 2 x 3 array
\describe{
 \item{survival}{dead, alive}
 \item{receptor.level}{amount of estrogen receptor that is present in breast tissue: low, high}
 \item{stage}{stage of disease: I, II, III}
}}
\source{
Newman (2001)
}
\usage{
breast.stage
}
\description{
A random sample of 199 female breast cancer patients who registered with the Northern Alberta Breast 
Cancer Registry in 1985. Entry into the cohort was restricted to women with either stage I, II, or III 
disease. Of the 199 subjects in the cohort, seven died of a cause other than breast cancer. These 
individuals were dropped from the analysis. Data are provided in different subsets and formats 
throughout Newman: Table 4.5(a), page 99 (\code{\link{breast}}); 
Table 5.3, page 126 (\code{\link{breast.receptor}}); 
Table 5.10, page 140 (\code{\link{breast.stage}}); 
Table 9.1, page 175 (\code{\link{breast.survival}})
}
\keyword{datasets}

