% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZtests_BME.R
\name{getZtests_BME}
\alias{getZtests_BME}
\title{Get adaptive enrichment design test statistics for a given trial.}
\usage{
getZtests_BME(zstats, alpha1 = 0.0125, alpha2 = 0.0125)
}
\arguments{
\item{zstats}{An object returned by \link{getZstats_FA}}

\item{alpha1}{Significance level for subgroup S}

\item{alpha2}{Significance level for full population F}
}
\value{
A list indicating whether the null hypothesis is rejected in S and/or F
}
\description{
Get adaptive enrichment design test statistics for a given trial.
}
\examples{
d1 <- simu_enrich_trial(n = 600, prop_S = 0.5, duration = 20)
d2 <- simu_enrich_trial(n = 100, prop_S = 0.5, duration = 20)
zstats <- getZstats_FA(d1, d2, targetEvents = c(210, 280, 420), expand=0)
getZtests_BME(zstats, alpha1 = 0.0125, alpha2 = 0.0125)
}
