% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh-test.R
\name{mantelhaen.pt.test}
\alias{mantelhaen.pt.test}
\title{Mantel-Haenszel Test of association for person-time data}
\usage{
mantelhaen.pt.test(time, status, exposure)
}
\arguments{
\item{time}{a numeric vector of survival times.}

\item{status}{a numeric vector of censoring indicators, with 0 = censored and
1 = dead.}

\item{exposure}{a factor vector with two levels indicating exposure. The
first level is assumed to be the exposed condition.}
}
\value{
A list with class \code{"htest"} containing the following 
  components: 
  \describe{ 
   \item{statistic}{The Mantel-Haenszel X-squared chi-square test statistic.} 
   \item{p.value}{The p-value of the test.} 
   \item{estimate}{Hazard ratio estimates. The reference level is 1.} 
   \item{null.value}{The null hazard ratio, which is currently set to 1.} 
   \item{alternative}{A character string describing the alternative hypothesis. Currently only "two.sided".} 
   \item{method}{A character string indicating the method employed.} 
   \item{data.name}{A character string giving the name of the data.} 
  }
}
\description{
Performs a Mantel-Haenszel test of the null that the hazard ratio is 1 for 
all exposure categories, relative to the reference category. A warning is 
given if any expected count is less than 5, along with the offending expected
count(s).
}
\examples{
## Example 10.12
with(breast.survival, mantelhaen.pt.test(time, status, stage))

## Example 10.13, assessment of the Poisson-Exponential Assumption
mantelhaen.pt.test(time = c(2363, 7108), status = c(5,44), exposure = gl(n = 2, k = 1))
## p = 0.017; moderate evidence of unequal hazard rates in the two time intervals.
}
\references{
Newman (2001), page 216.
}

